requirements.txt
appdirs==1.4.3
bson==0.4.7
cycler==0.10.0
matplotlib==2.0.0
numpy==1.12.1
olefile==0.44
packaging==16.8
Pillow==4.1.0
pkg-resources==0.0.0
pymongo==3.4.0
pyparsing==2.2.0
PyQt5==5.8.2
python-dateutil==2.6.0
pytz==2017.2
scikit-learn==0.18.1
scipy==0.19.0
sip==4.19.2
six==1.10.0
